requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-stdimage==5.3.0
Pillow==9.0.1
python-decouple==3.5
sqlparse==0.4.2
